refactor: colocate tests with source files per Svelte convention
