+++ src/lib/utils/escapeRegExp.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { escapeRegExp } from '../escapeRegExp.js';
+import { escapeRegExp } from './escapeRegExp.js';
 
 describe('escapeRegExp', () => {
 	it('should escape special regex characters', () => {
